+++ noodles-sam/src/record/mapping_quality.rs
@@ -1,6 +1,6 @@
 //! SAM record mapping quality.
 
-use std::ops::Deref;
+use std::{error, fmt, num, ops::Deref, str::FromStr};
 
 /// The raw value of a missing mapping quality.
 pub const MISSING: u8 = 255;
@@ -20,6 +20,32 @@ impl From<u8> for MappingQuality {
         } else {
             Self(Some(n))
         }
+    }
+}
+
+/// An error returned when a raw SAM record mapping quality fails to parse.
+#[derive(Clone, Debug, Eq, PartialEq)]
+pub enum ParseError {
+    /// The input failed to parse as an integer.
+    Parse(num::ParseIntError),
+}
+
+impl error::Error for ParseError {}
+
+impl fmt::Display for ParseError {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        match self {
+            Self::Parse(e) => write!(f, "parse error: {}", e),
+        }
+    }
+}
+
+impl FromStr for MappingQuality {
+    type Err = ParseError;
+
+    fn from_str(s: &str) -> Result<Self, Self::Err> {
+        let n: u8 = s.parse().map_err(ParseError::Parse)?;
+        Ok(Self::from(n))
     }
 }
 
@@ -48,6 +74,24 @@ mod tests {
         assert_eq!(*MappingQuality::from(13), Some(13));
         assert_eq!(*MappingQuality::from(144), Some(144));
         assert_eq!(*MappingQuality::from(255), None);
+    }
+
+    #[test]
+    fn test_from_str() {
+        assert_eq!("0".parse(), Ok(MappingQuality::from(0)));
+        assert_eq!("8".parse(), Ok(MappingQuality::from(8)));
+        assert_eq!("13".parse(), Ok(MappingQuality::from(13)));
+        assert_eq!("144".parse(), Ok(MappingQuality::from(144)));
+        assert_eq!("255".parse(), Ok(MappingQuality::from(255)));
+
+        assert!(matches!(
+            "".parse::<MappingQuality>(),
+            Err(ParseError::Parse(_))
+        ));
+        assert!(matches!(
+            "256".parse::<MappingQuality>(),
+            Err(ParseError::Parse(_))
+        ));
     }
 
     #[test]